refactor: extract component out of other component and put it in a separate file
fixup
