+++ src/features/identityCheck/components/countryPicker/formatCallingCode.ts
@@ -1,0 +1,1 @@
+export const formatCallingCode = (code: string) => `+${code}`
